Add new useDateRangeSelectValue composition
