+++ src/compositions/useWorkspaceDashboardFilter.ts
@@ -1,41 +1,19 @@
 import { DateRangeSelectValue } from '@prefecthq/prefect-design'
-import { useRouteQueryParam, DateRouteParam, NumberRouteParam } from '@prefecthq/vue-compositions'
+import { useRouteQueryParam } from '@prefecthq/vue-compositions'
 import { computed, reactive } from 'vue'
+import { useDateRangeSelectValueFromRoute } from '@/compositions/useDateRangeSelectValue'
 import { WorkspaceDashboardFilter } from '@/types/dashboard'
 
 export function useWorkspaceDashboardFilterFromRoute(defaultValue: WorkspaceDashboardFilter): WorkspaceDashboardFilter {
-  const startDate = useRouteQueryParam('startDate', DateRouteParam)
-  const endDate = useRouteQueryParam('endDate', DateRouteParam)
-  const seconds = useRouteQueryParam('seconds', NumberRouteParam)
-  const type = useRouteQueryParam('type')
   const tags = useRouteQueryParam('tags', defaultValue.tags)
+  const dateRange = useDateRangeSelectValueFromRoute(defaultValue.range)
 
   const range = computed<NonNullable<DateRangeSelectValue>>({
     get() {
-      if (type.value === 'range' && startDate.value && endDate.value) {
-        return { type: 'range', startDate: startDate.value, endDate: endDate.value }
-      }
-
-      if (type.value === 'span' && seconds.value) {
-        return { type: 'span', seconds: seconds.value }
-      }
-
-      return defaultValue.range
+      return dateRange.range ?? defaultValue.range
     },
     set(value) {
-      if (value.type === 'range') {
-        type.value = 'range'
-        startDate.value = value.startDate
-        endDate.value = value.endDate
-        seconds.value = undefined
-      }
-
-      if (value.type === 'span') {
-        type.value = 'span'
-        startDate.value = undefined
-        endDate.value = undefined
-        seconds.value = value.seconds
-      }
+      dateRange.range = value
     },
   })
